+++ example/lib/pages/player/rotation/human_topdown_player.dart
@@ -1,0 +1,14 @@
+import 'package:bonfire/bonfire.dart';
+import 'package:example/pages/player/rotation/human_topdown_player_spritesheet.dart';
+
+class HumanTopdownPlayer extends RotationPlayer {
+  HumanTopdownPlayer({
+    required Vector2 position,
+  }) : super(
+          position: position,
+          size: Vector2(24, 16),
+          animIdle: HumanTopdownPlayerSpritesheet.getIdle(),
+          animRun: HumanTopdownPlayerSpritesheet.getRun(),
+          speed: 32,
+        );
+}