+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# Next
+- Fix `RotationPlayer` animation bug.
+- Add an option for using to enable diagonal input events on `KeyboardConfig` and `JoystickDirectional`.
+
+
 # 3.0.0-beta.15
 
 - `GameObject` improvements
@@ -5,7 +10,7 @@
 - Adds `initPosition` param in `CameraConfig`
 - Adds `replacePlatformAnimation` method in `PlatformPlayer`
 - Collision system improvements.
-- Add an option for using to enable diagonal input events on `KeyboardConfig` and `JoystickDirectional`.
+
 
 # 3.0.0-beta.14